BS-10 : Make the grid as a responsive
