+++ app/bidding-dashboard/page.tsx
@@ -1,5 +1,5 @@
 import React from 'react'
-import OfferTable from '@/components/OfferTable'
+import OfferTable from '@/components/offer-table/OfferTable'
 
 const page = () => {
   return (
